requirements.txt
#kokoro -- not working
piper-tts